+++ v1/models/protection_policy_request.go
@@ -124,7 +124,7 @@ type ProtectionPolicyRequest struct {
 
 	// Specifies the number of mins/hours/days in seconds to retain
 	// CDP backups if CDP schedule exists.
-	NumSecsToKeep *int32 `json:"numSecsToKeep,omitempty"`
+	NumSecsToKeep *int64 `json:"numSecsToKeep,omitempty"`
 
 	// Specifies the parent global policy Id. This must be specified when
 	// creating a policy from global policy template.
